sistema_de_informes/services/rest-api/requirements.txt
fastapi==0.111.0
uvicorn==0.30.0
pydantic==2.8.2
SQLAlchemy==2.0.32
passlib[bcrypt]==1.7.4
python-jose==3.3.0
python-multipart==0.0.9
